// src/sap/dto/sap-transaction-response.dto.ts
export class SapTransactionResponseDto {
  id: string;
  cardId: string;
  customerId: string;
  date: string;
  time: string;
  description: string;
  merchantName: string;
  merchantCategory: string;
  amount: number;
  currency: string;
  type: string;
  status: string;
  authCode: string;
  country: string;
}